text classification/requirements.txt
numpy~=1.19.2
tensorflow-gpu~=1.13.1